[pull request] cleaning up some final comments
This should be the last commit before merging to development.

The implementation includes the initial basis for the /post endpoint to create crowdactions from the CMS.

The continuation of this task (/post) should be done ideally in the CMS task (up for a discussion, of course) but could expanded somewhere else for other purposes.

For example, the badge issue may implement a very rudimentary version of this /post crowdaction for testing purposes.